Better Exception handling by javax.bluetooth and javax.microedition.location classes.
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@2627 06c8f799-c779-4783-8af7-179afedb67c8

+++ classes/lejos/keyboard/Keyboard.java
@@ -375,7 +375,7 @@ public class Keyboard extends Thread {
 			id = KeyEvent.KEY_RELEASED;
 		}
 		
-		int code = getJavaConstant((byte)normalizedScanCode);
+		int code = getJavaConstant(normalizedScanCode);
 		
 		// Recalculate modifier
 		recalculateModifier(code, id);
